predict_celltype.Seurat performance improvement
predict_celltype.Seurat was converting the full data matrix from
sparse to dense. Then the predict_celltype.matrix method filtered
only the genes present in the db. This is not optimal as it
generates a potentially huge matrix to later on keep a much smaller
number of rows.

Instead, a method for predicting sparse matrices is given, that
first filters the genes and then converts the matrix from sparse
to dense.